+++ tests/org.eclipse.ocl.examples.xtext.tests/src/org/eclipse/ocl/examples/test/xtext/PrecedenceTests.java
@@ -53,12 +53,12 @@ public class PrecedenceTests extends XtextTestCase
 		Precedence p2c = createPrecedence(library2, "D", AssociativityKind.LEFT);
 		libraries.add(library2);
 		List<String> errors = new PrecedenceManager().compilePrecedences(libraries);
-		assertEquals(0, p1a.getOrder().intValue());
-		assertEquals(1, p1b.getOrder().intValue());
-		assertEquals(3, p1c.getOrder().intValue());
-		assertEquals(1, p2a.getOrder().intValue());
-		assertEquals(2, p2b.getOrder().intValue());
-		assertEquals(3, p2c.getOrder().intValue());
+		assertEquals(0, p1a.getOrder());
+		assertEquals(1, p1b.getOrder());
+		assertEquals(3, p1c.getOrder());
+		assertEquals(1, p2a.getOrder());
+		assertEquals(2, p2b.getOrder());
+		assertEquals(3, p2c.getOrder());
 		assertEquals(0, errors.size());
 	}
 	
@@ -73,10 +73,10 @@ public class PrecedenceTests extends XtextTestCase
 		Precedence p2b = createPrecedence(library2, "A", AssociativityKind.LEFT);
 		libraries.add(library2);
 		List<String> errors = new PrecedenceManager().compilePrecedences(libraries);
-		assertEquals(0, p1a.getOrder().intValue());
-		assertEquals(1, p1b.getOrder().intValue());
-		assertEquals(1, p2a.getOrder().intValue());
-		assertEquals(0, p2b.getOrder().intValue());
+		assertEquals(0, p1a.getOrder());
+		assertEquals(1, p1b.getOrder());
+		assertEquals(1, p2a.getOrder());
+		assertEquals(0, p2b.getOrder());
 		assertEquals(1, errors.size());
 	}
 	
@@ -89,8 +89,8 @@ public class PrecedenceTests extends XtextTestCase
 		Precedence p2a = createPrecedence(library2, "A", AssociativityKind.RIGHT);
 		libraries.add(library2);
 		List<String> errors = new PrecedenceManager().compilePrecedences(libraries);
-		assertEquals(0, p1a.getOrder().intValue());
-		assertEquals(0, p2a.getOrder().intValue());
+		assertEquals(0, p1a.getOrder());
+		assertEquals(0, p2a.getOrder());
 		assertEquals(1, errors.size());
 	}
 	
@@ -107,12 +107,12 @@ public class PrecedenceTests extends XtextTestCase
 		Precedence p2c = createPrecedence(library2, "D", AssociativityKind.LEFT);
 		libraries.add(library2);
 		List<String> errors = new PrecedenceManager().compilePrecedences(libraries);
-		assertEquals(0, p1a.getOrder().intValue());
-		assertEquals(2, p1b.getOrder().intValue());
-		assertEquals(3, p1c.getOrder().intValue());
-		assertEquals(0, p2a.getOrder().intValue());
-		assertEquals(1, p2b.getOrder().intValue());
-		assertEquals(3, p2c.getOrder().intValue());
+		assertEquals(0, p1a.getOrder());
+		assertEquals(2, p1b.getOrder());
+		assertEquals(3, p1c.getOrder());
+		assertEquals(0, p2a.getOrder());
+		assertEquals(1, p2b.getOrder());
+		assertEquals(3, p2c.getOrder());
 		assertEquals(1, errors.size());
 	}
 	
@@ -127,10 +127,10 @@ public class PrecedenceTests extends XtextTestCase
 		Precedence p2b = createPrecedence(library2, "C", AssociativityKind.LEFT);
 		libraries.add(library2);
 		List<String> errors = new PrecedenceManager().compilePrecedences(libraries);
-		assertEquals(0, p1a.getOrder().intValue());
-		assertEquals(2, p1b.getOrder().intValue());
-		assertEquals(0, p2a.getOrder().intValue());
-		assertEquals(1, p2b.getOrder().intValue());
+		assertEquals(0, p1a.getOrder());
+		assertEquals(2, p1b.getOrder());
+		assertEquals(0, p2a.getOrder());
+		assertEquals(1, p2b.getOrder());
 		assertEquals(1, errors.size());
 	}
 }
